refactor: biome 중복 제거
