+++ backend/pokerogue/src/main/java/com/pokerogue/helper/pokemon2/config/Pokemon2DatabaseInitializer.java
@@ -1,5 +1,8 @@
 package com.pokerogue.helper.pokemon2.config;
 
+import com.pokerogue.helper.ability2.data.Ability;
+import com.pokerogue.helper.ability2.data.AbilityInfo;
+import com.pokerogue.helper.ability2.repository.AbilityRepository;
 import com.pokerogue.helper.pokemon2.data.Evolution;
 import com.pokerogue.helper.pokemon2.data.EvolutionChain;
 import com.pokerogue.helper.pokemon2.data.Move;
@@ -24,11 +27,12 @@ import java.util.StringTokenizer;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.stereotype.Component;
 
-
+@Slf4j
 @Component
 @RequiredArgsConstructor
 public class Pokemon2DatabaseInitializer implements ApplicationRunner {
@@ -36,6 +40,7 @@ public class Pokemon2DatabaseInitializer implements ApplicationRunner {
     private final Pokemon2Repository pokemon2Repository;
     private final MoveRepository moveRepository;
     private final EvolutionRepository evolutionRepository;
+    private final AbilityRepository abilityRepository;
     private final List<String> pokemonKeys = List.of(
             "id",
             "speciesId",
@@ -85,6 +90,42 @@ public class Pokemon2DatabaseInitializer implements ApplicationRunner {
         save("pokemon.txt", this::savePokemon);
         save("move-for-pokemon-response.txt", this::saveMove);
         save("evolution-for-pokemon-response.txt", this::saveEvolution);
+        List<AbilityInfo> abilityInfos = new ArrayList<>();
+
+        try (InputStream inputStream = getClass().getClassLoader().getResourceAsStream("ability.txt");
+             BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(inputStream))) {
+            while (true) {
+                String abilityInfo = bufferedReader.readLine();
+                if (abilityInfo == null) {
+                    break;
+                }
+                abilityInfos.add(new AbilityInfo(abilityInfo));
+            }
+        } catch (IOException e) {
+            log.error("error message : {}", e.getStackTrace()[0]);
+        }
+
+        abilityInfos.stream()
+                .map(abilityInfo -> new Ability(
+                        abilityInfo.getId(),
+                        abilityInfo.getName(),
+                        abilityInfo.getDescription(),
+                        getAbilityPokemon(abilityInfo.getPokemons()))
+                ).forEach(abilityRepository::save);
+    }
+
+    private List<Pokemon> getAbilityPokemon(List<String> pokemons) {
+        List<Pokemon> abilityPokemons = new ArrayList<>();
+        for (int i = 0; i < pokemons.size(); i++) {
+            String pokemonId = pokemons.get(i);
+
+            Pokemon pokemon = pokemon2Repository.findById(pokemonId)
+                    .orElseThrow();
+
+            abilityPokemons.add(pokemon);
+        }
+
+        return abilityPokemons;
     }
 
     private void save(String file, Consumer<BufferedReader> consumer) {
