--- backend/pokerogue/src/main/java/com/pokerogue/helper/pokemon2/dto/BiomeResponse.java
@@ -1,5 +1,0 @@
-package com.pokerogue.helper.pokemon2.dto;
-
-public record BiomeResponse(String id, String name, String image) {
-
-}
